+++ ProesBack/Domain/Interfaces/ILoginRepository.cs
@@ -4,11 +4,11 @@ namespace ProesBack.Domain.Interfaces
 {
     public interface ILoginRepository
     {
-        Login Get(int id);
+        Login Get(long id);
         Login Login(string username, string password);
         void Insert(Login login);
         void Update(Login login);
-        void Delete(int id);
+        void Delete(long id);
         void Save();
     }
 }
